move VM sleep code to vm dir
